requirements.txt
Flask
Werkzeug
Flask-Cors
Pillow
google-generativeai
python-dotenv
Gunicorn
Waitress